Left and Right Functionality
Changed around some of the pins so i could connect everything easier. Also added better pwm control from the L928n motor driver. Was having a weird noise issue with motors and esp32 when using the same voltage source, but got around it by powering the esp32 with the 5v out on the Motor Driver Board. Made deadzones slightly bigger so it is easier to go from forward to left.